chore: catalog_data and status
